Adding core functionnalities You can now review a specific news by clicking on it : This will display a new window with the full article.
We also added some animations to the application (when navigating
from the list to an article) and made some design improvements.

Further design improvements are coming, such as adding pictures
above the articles.

The next big step is to get notifications to work and to add
links to give the user the ability to send a feedback.